feat: Prevent past bin days from being stored
This commit fixes a bug where bin collection days from the past could be stored and displayed to the user.

The filtering logic has been added to the `setBinDays` method in the `GlobalStateNotifier`, which ensures that no past dates are ever stored in the app's state.

An extension method `isTodayOrAfter` was added to `DateTime` to facilitate the filtering. This method was also simplified to be more concise.

The app version has been bumped to 2.2.4+1 and the changelog has been updated.